\name{coefLower}
\alias{coefLower}
\alias{coefLower<-}
\alias{coefUpper}
\alias{coefUpper<-}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Extract or Set Lower/Upper Bounds on Coefficients
}
\description{

  Extract or set lower/upper bounds on coefficients for covariance
  kernel objects.
  
}
\usage{
coefLower(object, ...)
coefUpper(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{

    A covariance kernel object.

  }
  \item{\dots}{

    Other arguments for methods.

  }
}
%% \details{}
\value{

  The lower or upper bounds on the covariance kernel parameters.
  
}
%% \references{}
%% \author{ODY}
%% \note{}
%% \seealso{}
%%  \examples{ }
%% \keyword{ ~kwd1 }
%% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
