\name{coef<-}
\alias{coef<-}
\title{
  Generic Function: Replacement of Coefficient Values
}
\description{
  
  Generic function for the replacement of coefficient values.
  
}
\usage{
`coef<-`(object, ..., value)
}
\arguments{
  \item{object}{
    
    Object having a numeric vector of coefficients, typically a
    covariance kernel object.
    
  }
  \item{\dots}{
    
    Other arguments for methods.

  }
  \item{value}{
    
    The value of the coefficients to be set.
    
  }
}

\value{
  
  The modified object.
  
}

%% \examples{ }
%%\keyword{ ~kwd1 }
%%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
