\name{gls}
\alias{gls}
\title{
  
  Generic Function: Generalized Least Squares Estimation
  with a Given Covariance Kernel

}
\description{
  
  Generic function computing a Generalized Least Squares estimation
  with a given covariance kernel.

}

\usage{
gls(object, ...)
}

\arguments{
  \item{object}{

    An object representing a covariance kernel.

  }
  \item{\dots}{

    Other arguments for methods.

  }
}
\value{
  
  A list with several elements corresponding to the estimation results.

}

%% \author{Yves Deville, Olivier Roustant}

