\name{kernelName}
\alias{kernelName}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Name of the One-Dimensional Kernel in a Composite Kernel Object
}
\description{
  Name of the 1d kernel in a composite kernel object. 
}
\usage{
kernelName(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
    A covariance kernel.
  }
  \item{\dots}{
    Arguments for methods.
  }
}
%\details{}
\value{
  A character string giving the kernel name.
}
%% \references{}
%% \author{}
%% \note{}
%% ~Make other sections like Warning with \section{Warning }{....} ~
%%\seealso{}
%% \examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
