\name{npar-methods}
\docType{methods}
%% \alias{npar-methods}
\alias{npar,covMan-method}
\alias{npar,covTS-method}
\alias{npar,covMan-method}
\title{
  
  Number of Parameters for a Covariance Kernel Object
  
}
\description{
  
  Number of parameters for a covariance kernel object.

}
\usage{

\S4method{npar}{covMan}(object, ...)

\S4method{npar}{covTS}(object, ...)

}
\arguments{
  \item{object}{

    An object with S4 class corresponding to a covariance kernel.
    
  }
  \item{...}{

    Not used yet.

  } 
}
\value{
   
  The number of parameters.

}
%%\references{}
%%\author{}
%%\section{Caution}{}
\seealso{

  \code{\link{coef}} method

}
%% \examples{}
%%\keyword{ ~kwd1 }
%%\keyword{ ~kwd2 }
