\name{npar}
\alias{npar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Generic function: Number of Free Parameters in a Covariance Kernel
}
\description{
  
  Generic function returning the number of free parameters in a
  covariance kernel.

}
\usage{
npar(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
    
    A covariance kernel object.

  }
  \item{\dots}{

    Other arguments for methods.

  }
}
%% \details{}
\value{

  The number of parameters.

}
%% \references{}
%% \author{}
%% \note{}
%% \seealso{}
%% \examples{}
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
