\name{parMap}
\alias{parMap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{

  Generic Function: Map the Parameters of a Composite Covariance
  Kernel

}
\description{

  Map the parameter of a composite covariance kernel on the inputs and
  the parameters of the 1d kernel.

}
\usage{
parMap(object, ...)
}
\arguments{
  \item{object}{
    
    A composite covariance kernel.

  }
  \item{\dots}{

    Arguments for methods.

  }
}
\value{

  A matrix with one row by input and one column for each of the
  parameters of the 1d kernel.

}
%% \references{}
%%\author{ }
%%\note{}
%%\seealso{}
%% \examples{}
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
