% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{parNamesSymm}
\alias{parNamesSymm}
\title{Vector of Names for the General 'Symm' Parameterisation}
\usage{
parNamesSymm(nlev)
}
\arguments{
\item{nlev}{Number of levels.}
}
\value{
Character vector of names.
}
\description{
Vector of names for the general 'Symm' parameterisation.
}
\examples{
parNamesSymm(nlev = 4)
}
